Update documentation and move DefaultExclusionProvider. Should be squashed when merging
