+++ processor/src/test/java/org/mapstruct/ap/test/nestedbeans/exclusions/custom/CustomMappingExclusionProvider.java
@@ -19,9 +19,11 @@
 package org.mapstruct.ap.test.nestedbeans.exclusions.custom;
 
 // tag::documentation[]
-import javax.lang.model.element.Name;
 
-import org.mapstruct.ap.spi.DefaultMappingExclusionProvider;
+import java.util.regex.Pattern;
+import javax.lang.model.element.Name;
+import javax.lang.model.element.TypeElement;
+
 import org.mapstruct.ap.spi.MappingExclusionProvider;
 
 // end::documentation[]
@@ -29,16 +31,17 @@ import org.mapstruct.ap.spi.MappingExclusionProvider;
  * @author Filip Hrisafov
  */
 // tag::documentation[]
-public class CustomMappingExclusionProvider extends DefaultMappingExclusionProvider
-    implements MappingExclusionProvider {
+public class CustomMappingExclusionProvider implements MappingExclusionProvider {
+    private static final Pattern JAVA_JAVAX_PACKAGE = Pattern.compile( "^javax?\\..*" );
 
     @Override
-    protected boolean isFullyQualifiedNameExcluded(Name name) {
+    public boolean isExcluded(TypeElement typeElement) {
         // end::documentation[]
         //For some reason the eclipse compiler does not work when you try to do NestedTarget.class
         // tag::documentation[]
-        return super.isFullyQualifiedNameExcluded( name ) ||
-            name.toString().equals( "org.mapstruct.ap.test.nestedbeans.exclusions.custom.Target.NestedTarget" );
+        Name name = typeElement.getQualifiedName();
+        return name.length() != 0 && ( JAVA_JAVAX_PACKAGE.matcher( name ).matches() ||
+            name.toString().equals( "org.mapstruct.ap.test.nestedbeans.exclusions.custom.Target.NestedTarget" ) );
     }
 }
 // end::documentation[]
